feat: user route and middleware

+++ pkg/repositories/interfaces/user.go
@@ -7,7 +7,7 @@ import (
 type UserRepository interface {
 	Create(*entities.User) (int64, error)
 	FindAll() ([]entities.User, error)
-	FindByID(int) (*entities.User, error)
+	FindByID(int64) (*entities.User, error)
 	FindByEmail(string) (*entities.User, error)
 	GetRoleID(string) (*entities.UserRole, error)
 	Update(*entities.User) (*entities.User, error)
